移動平均算出を各グラフへ移動
